lurch_devman: respond synchronously on device stops
Similarly as in 0639b85.
Note, that there's a bug - when one client request a stop, and meanwhile
other client requests a stop too, the second one will be responded to
after the stop is done. The first client will timeout.

There should be check if the machine is not in stopping state.